Implement step-ordering command
DexExpectStepOrder takes a (variadic) sequence of step numbers. Dexter
expects to see each StepOrder line in the order specified by the arguments.
Several errors can be reported: a line is completely missing, a line was
seen but not as often as expected, and finally that the ordering is
incorrect.

Displaying ordering problems to the user is difficult, because there's way
to tell what the nature of the difference from the correct order is.
Consider patches: there are multiple diff algorithms, and they all make
mistakes now and again when trying to concisely represent what's changed.
Thus, this patch makes no attempt to display the different ordering nicely,
it just displays it.