Add pass-covering dexter tests
The idea behind these is to manually create a set of tests that have
reasonable coverage of program behaviours, to increase the probability
that we can detect faulty debug info (or changes that break it).
Categorising by LLVM pass and trying to trigger different transforms
is part of achieving diversity in behaviour.