Acquire and print lit metrics from each test
This uses the existing facility in lit of "metrics" that can be attached
to tests. To do this, we use a test-format class that knows about dexters
output, and reads the dexter score out of the results csv. Lit then prints
it to stdout.

This is slightly exotic, but it enables all the other things that lit does
to be used with dexter, such as parallel jobs and test reports.