Add tests to stimulate simplifycfg after inlining
These tests stimulate specific simplifycfg transformations after inlining. The
folder name for each test is the name of the targeted simplifycfg function.

Ideally we'd like to be able to make assertions about what the call stack should
look like in the debugger for theses tests. These tests will need to be updated
when we get that functionaliy in DExTer.