Add return codes for high level functions
The return code is meant to represet program success on a high level.
It is not meant to replace exceptions.

The return code will is used as the program exit status.

OK(0)     - Normal termination, positive result
_ERROR(1) - Abnormal termination (usually an unhandled exception)
FAIL(2)   - Normal termination, negative result

This change will allow us to use the exit status to represent
whether a DExTer test has passed or failed. This feature is required
for debuginfo tests integration.