Implement DexVerify and LTD
DexVerify is a dexter command which can be used to verify a proposition about the
program trace.

A proposition can be made up of atomic propositions (assertions about what
the program state) and propositions connected and operators. The operators can
be boolean or temporal. The terminology and theory behind this is explained in
the document LTL_word.md included with this commit.

The term LTD is used to describe the DexVerify subcommands which are based on
Linear Temporal Logic (LTL).