Composite operators use default operator string representation
Composite operators were previously printing "{name} is {underlying-formula}({
args})" which was confusing for complex models. They now print only "{name}({
args})". The full trace (when implemented) will show how the operators are
composed under the hood.

Improve LTD error messages

If the wrong number of args are given to any class inheriting from any of
the classes in OperatorTypes.py the derived class name is printed with the
error.
name is printed with the error.