Tidy up
Add checks for DextStepIter validity before dereferencing

Fix LTD After() docstring

The docstring had switched p and q when describing After's behaviour. This will
probably be a common error because of the prefix notation. This has been raised
as an issue and will be discussed in the future.

Add missing newline to CompositeOperators.py