Cope with variable depth directory trees (#29)
os.path.split will only split the basename of a path from the dirname,
it won't split the entire path. Two levels of test directories will thus
make path-separators turn up in file names.

Get around this with a generator that splits the path into components,
that are then joined with '_'. Use this file basename for writing the
dextir file too, as that risks having path separators in its name too.