Provide warning if command is not immediately proceeded by '('
Without this patch, parsing 'command_name (' will cause an exception.

This change means the offending command will be ignored and a helpful warning
will be given.

Doing this instead of just ignoring the whitespace lets us refer to commands
in a test file without the command being parsed.

Note: Warnings can be suppresed with -w.