Add feature_tests/subtools (#47)
* Add feature_tests/commands

feature_tests are tests to be run with llvm-lit which verify that
dexter is working as expected. It is where regression tests will live.

feature_tests/commands test dexter commands in trivial scenarios to provide
a quick assurance that they're behaving correctly.

* Allow commands to be escaped in test files

Use `\` before a command to escape it:
    \DexLabel('this_is_never_evaluated')

* Add missing newline

* Add feature_tests/subtools

feature_tests are tests to be run with llvm-lit which verify that
dexter is working as expected. It is where regression tests will live.

feature_tests/subtools test dexter subtools in trivial scenarios to provide
a quick assurance that they're behaving correctly.