Set SUBSYSTEM:WINDOWS as the default environment on Windows
At the moment a console window is created when a DExTer test case is started up
for debugging on Windows. This is an undesirable default behaviour because the
test cases will steal focus every time they are run.

The default subsystem for programs with a 'main' entry point is CONSOLE.
Linking with /SUBSYSTEM:WINDOWS means the application will not create a console
when it runs.

/ENTRY:mainCRTStartup is now required because setting /SUBSYSTEM:WINDOWS tells
the linker to look for 'WinMain' by default instead of 'main'.

These compiler flags are overridable by providing them explicitly through
DExTer's --cflags and --ldflags.