Prevent FileExistsError when creating results and working directories
It is clear from the surrounding code for both cases that the intent is to only
create the directory if it doesn't already exist (i.e. using an existing
directory is okay). However, without this patch, if multiple dexter instances
try to create either directory simultaneously one instance will raise
FileExistsError.