Cope with debugged processes encountering exceptions
dbgeng will signal an exception has occurred if something like abort()
is called, however it seems to be difficult to detect this from the main
loop. Rather than terminating the debuggee, dbgeng just keeps trapping
on the same instruction, giving us an (almost) infinite loop.

Get around this by configuring what to do when an exception occurs,
adding a command to run when exceptions occur. That command causes
dbgeng to detach and quit; when we come to single step next time, if the
debuggee has gone away, execution completes.