Dbgeng fixes2 (#71)
* Correct reading of signed vals from dbgeng

* Correctly set is_irretrivable in dbgeng

* Cope with debugged processes encountering exceptions

dbgeng will signal an exception has occurred if something like abort()
is called, however it seems to be difficult to detect this from the main
loop. Rather than terminating the debuggee, dbgeng just keeps trapping
on the same instruction, giving us an (almost) infinite loop.

Get around this by configuring what to do when an exception occurs,
adding a command to run when exceptions occur. That command causes
dbgeng to detach and quit; when we come to single step next time, if the
debuggee has gone away, execution completes.

* Fix dbgeng evaluation of struct fields

It would appear that dbgeng reserves '.' for some special operation in
expressions that get evaluated. As a result, one can't write
"object.field", instead it will only accept "object->field". To ensure
that dexter presents a consistent input language, swap one with another
under the bonnet.

* Support expr eval in different frames in dbgeng